Clean up task registration
