+++ src/main/kotlin/com/bugsnag/android/gradle/internal/GradleUtil.kt
@@ -36,14 +36,6 @@ internal fun File.md5HashCode(): Int {
     }
 }
 
-internal fun <T: Task> TaskProvider<out T>.dependsOn(vararg tasks: TaskProvider<out Task>): TaskProvider<out T> {
-    if (tasks.isEmpty().not()) {
-        configure { it.dependsOn(*tasks) }
-    }
-
-    return this
-}
-
 /* Borrowed helper functions from the Gradle Kotlin DSL. */
 
 /**
